Refactor /web/PopupContainer.tsx
- Rename web/PopupContainer.tsx to web/PopupContainerView.tsx for consistency
with native-common.
- Pull the popup component tracking functionality in web/PopupContainerView.tsx
into a new base class common/PopupContainerViewBase.tsx.
